--- surface/launch
@@ -1,3 +1,0 @@
-cd ..
-python3 -m surface.launch "$@"
-cd surface
