+++ surface/xgui
@@ -1,0 +1,3 @@
+cd ..
+python3 -m surface.xgui "$@"
+cd surface
